# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediTech Smart Healthcare Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type HospitalType = (typeof HospitalType)[keyof typeof HospitalType];

export const HospitalType = {
  government: "government",
  private: "private",
  trust: "trust",
} as const;

export type HospitalStatus =
  (typeof HospitalStatus)[keyof typeof HospitalStatus];

export const HospitalStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface Hospital {
  id: number;
  name: string;
  address: string;
  city: string;
  state: string;
  phone: string;
  latitude: number;
  longitude: number;
  type: HospitalType;
  status: HospitalStatus;
  createdAt: string;
}

export interface HospitalResources {
  id: number;
  hospitalId: number;
  icuBeds: number;
  icuBedsAvailable: number;
  generalBeds: number;
  generalBedsAvailable: number;
  oxygenCylinders: number;
  oxygenCylindersAvailable: number;
  ambulancesTotal: number;
  ambulancesAvailable: number;
  doctorsOnDuty: number;
  ventilators: number;
  ventilatorsAvailable: number;
  updatedAt: string;
}

export interface UpdateHospitalResourcesBody {
  icuBedsAvailable?: number;
  generalBedsAvailable?: number;
  oxygenCylindersAvailable?: number;
  ambulancesAvailable?: number;
  doctorsOnDuty?: number;
  ventilatorsAvailable?: number;
}

export interface ResourcesSummary {
  totalHospitals: number;
  totalIcuBeds: number;
  availableIcuBeds: number;
  totalGeneralBeds: number;
  availableGeneralBeds: number;
  totalAmbulances: number;
  availableAmbulances: number;
  totalOxygen: number;
  availableOxygen: number;
  totalDoctors: number;
  activeBookings: number;
}

export type AmbulanceStatus =
  (typeof AmbulanceStatus)[keyof typeof AmbulanceStatus];

export const AmbulanceStatus = {
  available: "available",
  dispatched: "dispatched",
  maintenance: "maintenance",
} as const;

export type AmbulanceType = (typeof AmbulanceType)[keyof typeof AmbulanceType];

export const AmbulanceType = {
  basic: "basic",
  advanced: "advanced",
  icu: "icu",
} as const;

export interface Ambulance {
  id: number;
  hospitalId: number;
  vehicleNumber: string;
  driverName: string;
  driverPhone: string;
  status: AmbulanceStatus;
  type: AmbulanceType;
  latitude: number;
  longitude: number;
  updatedAt: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  dispatched: "dispatched",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type BookingEmergency =
  (typeof BookingEmergency)[keyof typeof BookingEmergency];

export const BookingEmergency = {
  critical: "critical",
  moderate: "moderate",
  low: "low",
} as const;

export interface Booking {
  id: number;
  patientName: string;
  patientPhone: string;
  pickupAddress: string;
  destinationHospitalId: number;
  ambulanceId?: number;
  status: BookingStatus;
  emergency: BookingEmergency;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateBookingBodyEmergency =
  (typeof CreateBookingBodyEmergency)[keyof typeof CreateBookingBodyEmergency];

export const CreateBookingBodyEmergency = {
  critical: "critical",
  moderate: "moderate",
  low: "low",
} as const;

export interface CreateBookingBody {
  patientName: string;
  patientPhone: string;
  pickupAddress: string;
  destinationHospitalId: number;
  emergency: CreateBookingBodyEmergency;
  notes?: string;
}

export type UpdateBookingStatusBodyStatus =
  (typeof UpdateBookingStatusBodyStatus)[keyof typeof UpdateBookingStatusBodyStatus];

export const UpdateBookingStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  dispatched: "dispatched",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateBookingStatusBody {
  status: UpdateBookingStatusBodyStatus;
}

export type EmergencyHotspotRiskLevel =
  (typeof EmergencyHotspotRiskLevel)[keyof typeof EmergencyHotspotRiskLevel];

export const EmergencyHotspotRiskLevel = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface EmergencyHotspot {
  id: number;
  area: string;
  city: string;
  latitude: number;
  longitude: number;
  riskLevel: EmergencyHotspotRiskLevel;
  incidentCount: number;
  predictedDemand: number;
  description: string;
}

export interface AnalyticsTrend {
  month: string;
  emergencyCalls: number;
  ambulancesDispatched: number;
  avgResponseTime: number;
  criticalCases: number;
}

export interface ResponseTimeStats {
  avgResponseTimeMinutes: number;
  bestResponseTimeMinutes: number;
  totalEmergencies: number;
  successRate: number;
  monthlyImprovement: number;
}

export type EmergencyAlertSeverity =
  (typeof EmergencyAlertSeverity)[keyof typeof EmergencyAlertSeverity];

export const EmergencyAlertSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface EmergencyAlert {
  id: number;
  title: string;
  message: string;
  severity: EmergencyAlertSeverity;
  hospitalId?: number;
  isActive: boolean;
  createdAt: string;
}

export type CreateAlertBodySeverity =
  (typeof CreateAlertBodySeverity)[keyof typeof CreateAlertBodySeverity];

export const CreateAlertBodySeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface CreateAlertBody {
  title: string;
  message: string;
  severity: CreateAlertBodySeverity;
  hospitalId?: number;
}

export interface DashboardSummary {
  totalHospitals: number;
  activeAmbulances: number;
  availableIcuBeds: number;
  availableGeneralBeds: number;
  pendingBookings: number;
  criticalAlerts: number;
  avgResponseTime: number;
  hospitalsAtCapacity: number;
  totalDoctorsOnDuty: number;
}

export type ListHospitalsParams = {
  city?: string;
  state?: string;
};

export type ListAmbulancesParams = {
  hospitalId?: number;
  status?: ListAmbulancesStatus;
};

export type ListAmbulancesStatus =
  (typeof ListAmbulancesStatus)[keyof typeof ListAmbulancesStatus];

export const ListAmbulancesStatus = {
  available: "available",
  dispatched: "dispatched",
  maintenance: "maintenance",
} as const;