# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediTech Smart Healthcare Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingEmergency } from "./bookingEmergency";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  patientName: string;
  patientPhone: string;
  pickupAddress: string;
  destinationHospitalId: number;
  ambulanceId?: number;
  status: BookingStatus;
  emergency: BookingEmergency;
  notes?: string;
  createdAt: Date;
  updatedAt: Date;
}