# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediTech Smart Healthcare Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalHospitals: number;
  activeAmbulances: number;
  availableIcuBeds: number;
  availableGeneralBeds: number;
  pendingBookings: number;
  criticalAlerts: number;
  avgResponseTime: number;
  hospitalsAtCapacity: number;
  totalDoctorsOnDuty: number;
}