# lib/api-zod/src/generated/types/hospitalResources.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediTech Smart Healthcare Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface HospitalResources {
  id: number;
  hospitalId: number;
  icuBeds: number;
  icuBedsAvailable: number;
  generalBeds: number;
  generalBedsAvailable: number;
  oxygenCylinders: number;
  oxygenCylindersAvailable: number;
  ambulancesTotal: number;
  ambulancesAvailable: number;
  doctorsOnDuty: number;
  ventilators: number;
  ventilatorsAvailable: number;
  updatedAt: Date;
}