# lib/api-zod/src/generated/types/resourcesSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediTech Smart Healthcare Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ResourcesSummary {
  totalHospitals: number;
  totalIcuBeds: number;
  availableIcuBeds: number;
  totalGeneralBeds: number;
  availableGeneralBeds: number;
  totalAmbulances: number;
  availableAmbulances: number;
  totalOxygen: number;
  availableOxygen: number;
  totalDoctors: number;
  activeBookings: number;
}