# lib/api-zod/src/generated/types/updateHospitalResourcesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MediTech Smart Healthcare Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateHospitalResourcesBody {
  icuBedsAvailable?: number;
  generalBedsAvailable?: number;
  oxygenCylindersAvailable?: number;
  ambulancesAvailable?: number;
  doctorsOnDuty?: number;
  ventilatorsAvailable?: number;
}